Add file check.

+++ orchestration/upload_dbt_file_to_snowflake.py
@@ -29,7 +29,7 @@ def get_file_name(config_name):
         return "target/manifest.json"
     elif config_name == "gdpr_logs":
         parse_log_data("logs/dbt.log", "log_data.json")
-        return "'log_data.json'"
+        return "log_data.json"
     else:
         return "target/run_results.json"
 
@@ -41,12 +41,15 @@ def parse_log_data(log_file_name: str, output_json_name: str):
     :param output_json_name: File name to write json
     :return:
     """
-    with open(log_file_name, "r") as file:
-        log_data = file.readlines()
+    if os.path.exists(log_file_name):
+        with open(log_file_name, "r") as file:
+            log_data = file.readlines()
 
-    parsed_data = [json.loads(line) for line in log_data]
-    with open(output_json_name, "w", encoding="utf-8") as f:
-        json.dump(parsed_data, f, ensure_ascii=False, indent=4)
+        parsed_data = [json.loads(line) for line in log_data]
+        with open(output_json_name, "w", encoding="utf-8") as f:
+            json.dump(parsed_data, f, ensure_ascii=False, indent=4)
+    else:
+        logging.warning(f"File not found: {log_file_name}")
 
 
 def get_table_name(config_name, snowflake_database):
